requirements.txt
asgiref==3.5.2
autopep8==2.0.0
backports.zoneinfo==0.2.1
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
pycodestyle==2.10.0
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
